[WIP] CompactionIteratorToInternalIterator support Seek

+++ db/compaction_picker.cc
@@ -1770,7 +1770,7 @@ Compaction* UniversalCompactionPicker::PickCompactionUniversalReadAmp(
   }
 
   uint64_t estimated_total_size = 0;
-  for (unsigned int i = start_index; i < first_index_after; i++) {
+  for (size_t i = start_index; i < first_index_after; i++) {
     estimated_total_size += sorted_runs[i].size;
   }
   uint32_t path_id = GetPathId(ioptions_, estimated_total_size);
